Ensure protected path starts with path name
Avoid issue where request path contains the name of a protected path,
but is not actually a protected path. For example, if the protected path
is `/downloads/`, a request to `/public/downloads/<file>` would also
match with the previous use of path.includes(). However, this may not
be intended.

This change replaces path.includes() with path.startsWith() to require
HTTP request paths to start with a protected path name in order to be
processed by this filter.